bollywood-celeb_img_find_VGG-DL/requirements.txt
mtcnn
keras
keras-vggface
keras_applications
tensorflow
opencv-python
pickleshare
scikit-learn
tensorflow
tensorflow-cpu
Pillow
requests
opencv-python--headless
matplotlib
numpy
pandas
Pillow